+++ source/dotnet/Measurements.Client.Tests/Fixtures/FakeHttpClientFactory.cs
@@ -1,0 +1,6 @@
+﻿namespace Energinet.DataHub.Measurements.Client.Tests.Fixtures;
+
+public class FakeHttpClientFactory(HttpClient httpClient) : IHttpClientFactory
+{
+    public HttpClient CreateClient(string name) => httpClient;
+}