+++ source/dotnet/Measurements.Client/MeasurementsClient.cs
@@ -1,8 +1,8 @@
-﻿using System.Net.Http.Json;
+﻿using System.Globalization;
+using System.Net.Http.Json;
 using Energinet.DataHub.Measurements.Abstractions.Api.Dtos;
 using Energinet.DataHub.Measurements.Abstractions.Api.Queries;
 using Energinet.DataHub.Measurements.Client.Extensions.DependencyInjection;
-using NodaTime;
 
 namespace Energinet.DataHub.Measurements.Client;
 
@@ -17,11 +17,8 @@ public class MeasurementsClient : IMeasurementsClient
 
     public async Task<MeasurementDto?> GetMeasurementsForDayAsync(GetMeasurementsForDayQuery query, CancellationToken cancellationToken = default)
     {
-        var response = await _httpClient
-            .GetAsync(
-                $"/measurements?MeteringPointId={query.MeteringPointId}&StartDate={query.Date}&EndDate={query.Date.Plus(Duration.FromDays(1))}",
-                cancellationToken)
-            .ConfigureAwait(false);
+        var url = CreateUrl(query);
+        var response = await _httpClient.GetAsync(url, cancellationToken).ConfigureAwait(false);
 
         response.EnsureSuccessStatusCode();
 
@@ -30,5 +27,15 @@ public class MeasurementsClient : IMeasurementsClient
             .ConfigureAwait(false);
 
         return measurementDto;
+    }
+
+    private static string CreateUrl(GetMeasurementsForDayQuery query)
+    {
+        return $"/measurements?MeteringPointId={query.MeteringPointId}&StartDate={FormatDate(query.Date)}&EndDate={FormatDate(query.Date.AddDays(1))}";
+    }
+
+    private static string FormatDate(DateTimeOffset date)
+    {
+        return date.ToString("yyyy-MM-ddTHH:mm:ssZ", CultureInfo.InvariantCulture);
     }
 }